init decoder training ecube
