+++ db/html/static/resources/js/list.js
@@ -443,7 +443,7 @@ TaskEntry.prototype.update = function(info) {
     this.files.show();
     this.files.update_filelist(info.datafiles, this.idx);
 
-    if (this.files.neural_data_found){
+    if (this.bmi && this.bmi._neuralinfo) {
         // Create the JS object to represent the BMI menu
         this.bmi = new BMI(this.idx, info.bmi, info.notes);
     }
